move most types from values to base
i have to
